--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -38,6 +38,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ## v0.9.0
 
 - Support all asset list v1 tokens
+- Use spot price in pricing
 
 ## v0.8.4