--- a/tokens/usecase/pricing/chain/pricing_chain.go
+++ b/tokens/usecase/pricing/chain/pricing_chain.go
@@ -110,7 +110,7 @@ func (c *chainPricing) GetPrice(ctx context.Context, baseDenom string, quoteDeno
 	}
 
 	// Create a quote denom coin.
-	oneQuoteCoin := sdk.NewCoin(quoteDenom, quoteDenomScalingFactor.TruncateInt())
+	tenQuoteCoin := sdk.NewCoin(quoteDenom, osmomath.NewInt(10).Mul(quoteDenomScalingFactor.TruncateInt()))
 
 	// Overwrite default config with custom values
 	// necessary for pricing.
@@ -122,20 +122,57 @@ func (c *chainPricing) GetPrice(ctx context.Context, baseDenom string, quoteDeno
 	routerConfig.MaxSplitIterations = domain.DisableSplitRoutes
 
 	// Compute a quote for one quote coin.
-	quote, err := c.RUsecase.GetOptimalQuoteFromConfig(ctx, oneQuoteCoin, baseDenom, routerConfig)
+	quote, err := c.RUsecase.GetOptimalQuoteFromConfig(ctx, tenQuoteCoin, baseDenom, routerConfig)
 	if err != nil {
 		return osmomath.BigDec{}, err
 	}
 
-	// Compute on-chain price for 1 unit of base denom and quote denom.
-	chainPrice := osmomath.NewBigDecFromBigInt(oneQuoteCoin.Amount.BigIntMut()).QuoMut(osmomath.NewBigDecFromBigInt(quote.GetAmountOut().BigIntMut()))
+	routes := quote.GetRoute()
+	if len(routes) == 0 {
+		return osmomath.BigDec{}, fmt.Errorf("no route found when computing pricing for %s (base) -> %s (quote)", baseDenom, quoteDenom)
+	}
+
+	route := routes[0]
+
+	chainPrice := osmomath.OneBigDec()
+
+	pools := route.GetPools()
+
+	tempQuoteDenom := quoteDenom
+	tempBaseDenom := baseDenom
+	useAlternativeMethod := false
+	for _, pool := range pools {
+
+		tempBaseDenom = pool.GetTokenOutDenom()
+
+		// Get spot price for the pool.
+		poolSpotPrice, err := c.RUsecase.GetPoolSpotPrice(ctx, pool.GetId(), tempQuoteDenom, tempBaseDenom)
+		if err != nil {
+			// TODO: alert
+			useAlternativeMethod = true
+			break
+		}
+
+		// Multiply spot price by the previous spot price.
+		chainPrice = chainPrice.MulMut(poolSpotPrice)
+
+		tempQuoteDenom = tempBaseDenom
+	}
+
+	if useAlternativeMethod {
+		// Compute on-chain price for 1 unit of base denom and quote denom.
+		chainPrice = osmomath.NewBigDecFromBigInt(tenQuoteCoin.Amount.BigIntMut()).QuoMut(osmomath.NewBigDecFromBigInt(quote.GetAmountOut().BigIntMut()))
+	} else {
+		chainPrice = osmomath.OneBigDec().QuoMut(chainPrice)
+	}
+
 	if chainPrice.IsZero() {
 		// Increase price truncation counter
 		pricesTruncationCounter.WithLabelValues(baseDenom, quoteDenom).Inc()
 	}
 
 	// Compute precision scaling factor.
-	precisionScalingFactor := osmomath.BigDecFromDec(baseDenomScalingFactor.Quo(oneQuoteCoin.Amount.ToLegacyDec()))
+	precisionScalingFactor := osmomath.BigDecFromDec(osmomath.NewDec(10).MulMut(baseDenomScalingFactor.Quo(tenQuoteCoin.Amount.ToLegacyDec())))
 
 	// Apply scaling facors to descale the amounts to real amounts.
 	currentPrice := chainPrice.MulMut(precisionScalingFactor)