## Benchmarks

### Pricing Logic


There isn't a meaningful change in performance
```
~/go/bin/benchstat old2 new2
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/sqs/tokens/usecase
cpu: Intel(R) Xeon(R) Gold 6248 CPU @ 2.50GHz
            │     old2     │             new2              │
            │    sec/op    │   sec/op     vs base          │
GetPrices-8   7.377m ± 27%   7.443m ± 3%  ~ (p=0.853 n=10)

            │     old2     │                new2                 │
            │     B/op     │     B/op      vs base               │
GetPrices-8   9.942Mi ± 0%   9.923Mi ± 0%  -0.20% (p=0.000 n=10)

            │    old2     │                new2                │
            │  allocs/op  │  allocs/op   vs base               │
GetPrices-8   138.2k ± 0%   137.8k ± 0%  -0.28% (p=0.000 n=10)
```

### Precompute Logic

Precompute
```
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/sqs/router/usecase
cpu: Intel(R) Xeon(R) Gold 6248 CPU @ 2.50GHz
BenchmarkGetPrecomputeOrderOfMagnitude-8   	 3527824	       371.8 ns/op	      96 B/op	       6 allocs/op
BenchmarkGetPrecomputeOrderOfMagnitude-8   	 3109318	       328.9 ns/op	      96 B/op	       6 allocs/op
BenchmarkGetPrecomputeOrderOfMagnitude-8   	 3569872	       349.1 ns/op	      96 B/op	       6 allocs/op
BenchmarkGetPrecomputeOrderOfMagnitude-8   	 3276049	       352.9 ns/op	      96 B/op	       6 allocs/op
BenchmarkGetPrecomputeOrderOfMagnitude-8   	 3739591	       322.9 ns/op	      96 B/op	       6 allocs/op
BenchmarkGetPrecomputeOrderOfMagnitude-8   	 3297483	       324.7 ns/op	      96 B/op	       6 allocs/op
PASS
ok  	github.com/osmosis-labs/sqs/router/usecase	10.494s
```

Non-precompute
```
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/sqs/router/usecase
cpu: Intel(R) Xeon(R) Gold 6248 CPU @ 2.50GHz
BenchmarkOrderOfMagnitude-8   	  229332	      5374 ns/op	    1256 B/op	      22 allocs/op
BenchmarkOrderOfMagnitude-8   	  239371	      5549 ns/op	    1256 B/op	      22 allocs/op
BenchmarkOrderOfMagnitude-8   	  257167	      5405 ns/op	    1256 B/op	      22 allocs/op
BenchmarkOrderOfMagnitude-8   	  236186	      4932 ns/op	    1256 B/op	      22 allocs/op
BenchmarkOrderOfMagnitude-8   	  233816	      5151 ns/op	    1256 B/op	      22 allocs/op
BenchmarkOrderOfMagnitude-8   	  222508	      8797 ns/op	    1256 B/op	      22 allocs/op
PASS
ok  	github.com/osmosis-labs/sqs/router/usecase	12.149s
```